requirements.txt
pyrogram
python-dotenv
tgcrypto
pycryptodomex
Pillow
python-magic
